Rename Herder -> HerderImpl, HerderGateway -> Herder.
