+++ Builds/VisualStudio2015/stellarx.vcxproj
@@ -151,7 +151,7 @@
     <ClCompile Include="..\..\src\crypto\SecretKey.cpp" />
     <ClCompile Include="..\..\src\database\Database.cpp" />
     <ClCompile Include="..\..\src\database\DatabaseTests.cpp" />
-    <ClCompile Include="..\..\src\herder\Herder.cpp" />
+    <ClCompile Include="..\..\src\herder\HerderImpl.cpp" />
     <ClCompile Include="..\..\src\herder\HerderTests.cpp" />
     <ClCompile Include="..\..\src\herder\TxSetFrame.cpp" />
     <ClCompile Include="..\..\src\history\CatchupStateMachine.cpp" />
@@ -290,8 +290,8 @@
     <ClInclude Include="..\..\src\crypto\SecretKey.h" />
     <ClInclude Include="..\..\src\database\Database.h" />
     <ClInclude Include="..\..\src\generated\StellarXDR.h" />
+    <ClInclude Include="..\..\src\herder\HerderImpl.h" />
     <ClInclude Include="..\..\src\herder\Herder.h" />
-    <ClInclude Include="..\..\src\herder\HerderGateway.h" />
     <ClInclude Include="..\..\src\herder\TxSetFrame.h" />
     <ClInclude Include="..\..\src\history\CatchupStateMachine.h" />
     <ClInclude Include="..\..\src\history\FileTransferInfo.h" />
